requirements_updated.txt
flask>=2.0.1
neo4j>=4.4.0
numpy>=1.25.0
pandas>=2.0.0
scikit-learn>=1.0.0
tensorflow-cpu>=2.13.0
scipy>=1.7.1
matplotlib>=3.4.3
seaborn>=0.11.2
werkzeug>=2.0.1
pytest>=6.2.5
